Support embedded images via a Content ID
Allows referencing attachments from the HTML body. This is used to
display images in the HTML email, where the image is an attachment
of the email. This solves the problem where Outlook desktop will not
display images if they need to be requested. However, it will display
them if they are part of the attachments.

This new feature requires upgrading ex_mail from 0.2.1 to 0.2.2. The
upgrade caused several tests to fail because of a behavior change in
the Mail.RFC2822.parse/1 function related to utf-8 handling. The new
behavior was converting ?utf-8? sequences to actual utf-8 characters.
Adjusting the tests to the new return values would have made the tests
less accurate in determining that values being sent to SES were correct.
The call to Mail.RFC2822.parse/1 was replaced with a new module that
returns values more like what the older Mail.RFC2822.parse/1 did.