Respect new arbitrary headers field for attachments
This change in contingent upon an update to the bamboo library to
include a `:headers` key in `Bamboo.Attachment` that holds a keyword
list of headers that will override *all* others.

For e.g., in the specific use-case of adding inlined attachments, some
special case headers are required. Now it's possible to add
`X-Attachment-Id` and override the Content-Disposition header with
`inline` to permit the proper display of inlined content, without an
email client showing them as explicit attachments for the user to
download.