Pass tags to + increase version of ex_aws_ses...
Now allows a `:tags` option which is respected by the latest version of
the `:ex_aws_ses` library.  This permits arbitary tags that will be
interpreted and kept with the email object on AWS.  As just one example,
this is useful for adding metadata that can be retrieved from failure
notifications sent to your app via SNS or Firehose.  Such metadata can
be important for any follow-up actions like unsubscribing a user or
cancelling in-flight jobs.

The library version has also been updated to `2.3.0` in the `mix.exs`.